+++ cli/commons/src/Microsoft.Kiota.Cli.Commons/IO/FormatterType.cs
@@ -1,0 +1,8 @@
+namespace Microsoft.Kiota.Cli.Commons.IO;
+
+public enum FormatterType
+{
+    JSON,
+    TABLE,
+    NONE
+}