+++ serialization/python/json/serialization_json/__init__.py
@@ -1,0 +1,6 @@
+"""
+Implementation of Kiota Serialization interfaces for JSON
+"""
+from ._version import VERSION
+
+__version__ = VERSION